Add --promote flag to gcloud app deploy command by default
The default behavior of the PROMOTE flag to the `flow gcappengine` is
true as stated in the README. However, the current behavior didn't add
the `--promote` flag to the built `gcloud app deploy` command.

This commit adds the `--promote` flag whenever it is set to true.

Also fix a documentation error in the README, the proper enviroment
variable is `GCAPPENGINE_USER_JSON` not `GCAPPENGINE_JSON_USER`.

Fixes [#19]